chore: rewrite json channel converters
