+++ src/Playwright/Transport/Converters/ChannelOwnerConverterFactory.cs
@@ -23,30 +23,49 @@
  */
 
 using System;
+using System.Runtime.InteropServices.ComTypes;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
 namespace Microsoft.Playwright.Transport.Converters;
 
-internal class ChannelToGuidConverter : JsonConverter<ChannelOwner>
+internal class ChannelOwnerConverterFactory : JsonConverterFactory
 {
     private readonly Connection _connection;
 
-    public ChannelToGuidConverter(Connection connection)
+    public ChannelOwnerConverterFactory(Connection connection)
     {
         _connection = connection;
     }
 
-    public override bool CanConvert(Type type) => typeof(ChannelOwner).IsAssignableFrom(type);
+    public override bool CanConvert(Type typeToConvert) => typeof(ChannelOwner).IsAssignableFrom(typeToConvert);
 
-    public override ChannelOwner Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    public override JsonConverter CreateConverter(Type typeToConvert, JsonSerializerOptions options)
+    {
+        return (JsonConverter)Activator.CreateInstance(
+            typeof(ChannelOwnerToGuidConverter<>).MakeGenericType(new Type[] { typeToConvert }),
+            new object[] { _connection });
+    }
+}
+
+internal class ChannelOwnerToGuidConverter<T> : JsonConverter<T>
+    where T : ChannelOwner
+{
+    private readonly Connection _connection;
+
+    public ChannelOwnerToGuidConverter(Connection connection)
+    {
+        _connection = connection;
+    }
+
+    public override T Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
         using JsonDocument document = JsonDocument.ParseValue(ref reader);
         string guid = document.RootElement.GetProperty("guid").ToString();
-        return _connection.GetObject(guid);
+        return _connection.GetObject(guid) as T;
     }
 
-    public override void Write(Utf8JsonWriter writer, ChannelOwner value, JsonSerializerOptions options)
+    public override void Write(Utf8JsonWriter writer, T value, JsonSerializerOptions options)
     {
         writer.WriteStartObject();
         writer.WriteString("guid", value.Guid);
