--- src/Playwright/Transport/Converters/ChannelOwnerToGuidConverter.cs
@@ -1,58 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) Microsoft Corporation.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-using System;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace Microsoft.Playwright.Transport.Converters;
-
-internal class ChannelOwnerToGuidConverter<T>
-    : JsonConverter<T>
-    where T : ChannelOwner
-{
-    private readonly Connection _connection;
-
-    public ChannelOwnerToGuidConverter(Connection connection)
-    {
-        _connection = connection;
-    }
-
-    public override bool CanConvert(Type type)
-        => (typeof(T) == typeof(ChannelOwner) && typeof(T).IsAssignableFrom(type)) || type == typeof(T);
-
-    public override T Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
-        using JsonDocument document = JsonDocument.ParseValue(ref reader);
-        string guid = document.RootElement.GetProperty("guid").ToString();
-        return _connection.GetObject(guid) as T;
-    }
-
-    public override void Write(Utf8JsonWriter writer, T value, JsonSerializerOptions options)
-    {
-        writer.WriteStartObject();
-        writer.WriteString("guid", value.Guid);
-        writer.WriteEndObject();
-    }
-}
